requirements.txt
# Core
python-dotenv
streamlit
asyncio

# PDF parsing
PyPDF2
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=0.0.8
google-generativeai>=0.4.0

# Vector store (FAISS)
faiss-cpu
tiktoken

# Optional: Useful utils for dev/debug
rich
